+++ components/api/__init__.py
@@ -1,0 +1,3 @@
+from details import *
+from elements import *
+from forms import *
